Prevent closing of ModalBottomSheet and SideSheet during emoji editing
The ModalBottomSheet for emoji editing can no longer be dismissed by the user.
The SideSheet is now configured to be non-modal when an emoji is being edited, preventing it from being closed by gestures or clicking outside.
A "Cancel" button has been added to the emoji editing SideSheet.
The "OK" and "Cancel" buttons in the SideSheet have been moved to the top.